backend/pyproject.toml
[tool.poetry]
name = "realtime-voice-rag-backend"
version = "0.1.0"
description = "Backend for VoiceRAG application"
authors = ["Devin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.3"}
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-multipart = "^0.0.16"
python-dotenv = "^1.0.1"
openai = "^1.52.2"
langchain = "^0.3.4"
chromadb = "^0.5.15"
tiktoken = "^0.8.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
websockets = "^13.1"
openai-whisper = "^20240930"
langchain-community = "^0.3.3"
langchain-openai = "^0.2.3"
sentence-transformers = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
